feat(web): display boundary shapefiles (idas-131)
- consolidated tests and deleted unnecessary files
